requirements.txt
dj-static==0.0.6
Django==1.11.1
django-bootstrap-form==3.2.1
django-heroku-memcacheify==1.0.0
django-pylibmc==0.6.1
django-pylibmc-sasl==0.2.4
gunicorn==19.7.1
honcho==0.6.6
newrelic==2.86.3.70
pylibmc==1.5.1
pystache==0.5.4
pytz==2017.2
requests==2.3.0
static==1.1.1
static3==0.5.1
